requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.134.0
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
websockets==12.0
pydantic==2.4.2
python-multipart==0.0.5